+++ spring-telqos-api/src/test/java/com/dwarfeng/springtelqos/api/launcher/Example.java
@@ -1,9 +1,15 @@
-package com.dwarfeng.springtelqos.impl.service;
+package com.dwarfeng.springtelqos.api.launcher;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.util.Scanner;
 
+/**
+ * Telqos 示例。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
 public class Example {
 
     public static void main(String[] args) {
@@ -17,8 +23,9 @@ public class Example {
 
         // 显示欢迎信息。
         System.out.println("开发者您好!");
-        System.out.println("这是一个示例, 用于演示 dwarfeng-ftp 的 Telqos 功能");
+        System.out.println("这是一个示例, 用于演示 spring-telqos-api 的 Telqos 集成功能");
         System.out.println("您可以使用 telnet 客户端工具访问本机 ${telqos.port} 端口以体验本示例的功能");
+        System.out.println("您可以测试 dubbo、memory、log4j2、uptime、jmx-remote 等命令");
         System.out.print("本示例的功能体验完成后，请按回车键结束示例...");
         scanner.nextLine();