+++ CHANGELOG.md
@@ -12,7 +12,10 @@
 
 ### 功能移除
 
-- (无)
+- 移除项目与 `spring-terminator` 的集成。
+
+- 移除不需要的依赖。
+  - 移除 `spring-terminator` 依赖。
 
 ---